o Added a constant (SELECT_TIMEOU° to replace the magic numbers used in the Processor o The select(int) is now a select(long), as the Selector.select() method uses a long, not an int. o Minor refactoring
git-svn-id: https://svn.apache.org/repos/asf/mina/trunk@751744 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mina/transport/socket/nio/NioProcessor.java
@@ -65,7 +65,7 @@ public final class NioProcessor extends AbstractPollingIoProcessor<NioSession> {
     }
 
     @Override
-    protected int select(int timeout) throws Exception {
+    protected int select(long timeout) throws Exception {
         return selector.select(timeout);
     }
